pyproject.toml
[project]
name = "sudoplayer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.12.13",
    "asyncio>=3.4.3",
    "discord-py>=2.5.2",
    "python-dotenv>=1.1.0",
    "redis[hiredis]>=6.2.0",
]
